LayoutTests/http/tests/websocket/tests/hixie76/long-invalid-header-expected.txt

CONSOLE MESSAGE: line 0: Unexpected CR in name at pppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppp…
Make sure WebSocket gives errors on long invalid upgrade header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket is closed
PASS successfullyParsed is true

TEST COMPLETE
